+++ dataland-frontend/src/components/ui/TaxoCard.vue
@@ -1,23 +1,25 @@
 <template>
-  <Card class="bg-gray-400">
+  <Card class="bg-white d-card mr-2">
     <template #title></template>
     <template #content>
       <div class="grid">
         <div class="col col-6 text-left">
           <strong>{{ title }}</strong>
         </div>
-        <div class="col col-6 text-right">
-          <strong class="text-green-500 text-xl"> {{ percentCalculation }} %</strong>
+        <div class="col col-6 text-right text-green-500">
+          <span class="font-semibold text-xl">{{ percentCalculation }}</span>
+          <span>%</span>
         </div>
       </div>
-      <ProgressBar :value="percentCalculation" :showValue="false" class="bg-black-alpha-20">
+      <ProgressBar :value="percentCalculation" :showValue="false" class="bg-black-alpha-20 d-progressbar">
       </ProgressBar>
-      <div class="grid">
-        <div class="col col-6 text-left">
-          <strong class="text-xl"> € {{ amount }} m</strong>
+      <div class="grid mt-4 ">
+        <div class="col col-6 text-left p-0 pl-2">
+          <span class="font-semibold text-lg">€ </span>
+          <span class="font-bold text-2xl">{{ amount }}</span>
+
+          <span class="font-semibold text-lg"> m</span>
           <p class="left-align"><strong>Out of total of € {{ total }} m</strong></p>
-        </div>
-        <div class="col col-6 text-right">
         </div>
       </div>
     </template>
@@ -58,3 +60,23 @@ export default {
   }
 }
 </script>
+<style>
+.d-progressbar {
+  height: 0.25rem;
+  border-radius: 0.25rem;
+}
+
+.d-card {
+  border-radius: 0.3rem;
+  box-shadow: 0px 0px 32px 8px rgba(30, 30, 31, 0.08);
+}
+
+.d-card > .p-card-body {
+  padding: 0 0.5rem 0 0.5rem;
+}
+
+.d-card > .p-card-body > .p-card-content {
+  padding: 0;
+}
+
+</style>
