+++ dataland-frontend/src/router/index.ts
@@ -2,7 +2,7 @@ import { createWebHistory, createRouter } from "vue-router";
 import CreateData from "@/components/forms/CreateData.vue";
 import SearchData from "@/components/forms/SearchData.vue";
 import WelcomeDataland from "@/components/WelcomeDataland.vue";
-import CompanyEU from "@/components/pages/taxonomy/TaxonomyData.vue";
+import CompanyEU from "@/components/pages/taxonomy/TaxonomyPanel.vue";
 import CompanyInformation from "@/components/pages/company/CompanyInformation.vue";
 import DataList from "@/components/pages/data/DataList.vue";
 import SearchTaxonomy from "@/components/pages/taxonomy/SearchTaxonomy.vue";
